* Follow to the changes in xAODTrigL1Calo-00-00-55:     * CMXCPHits_v1, CMXJetHits_v1: change size of hits0 and hits1 from uint8_t to uint32_t       (it happens that these fields can store extra information not only thresholds)     * CPMTower_v2, CMXCPTob_v1, CMXEtSums_v1, JetElement_v2, CMXJetTob_v1, CMXCPHits_v1:       - change errors size from uint8_t to uint32_t (for unified usage in our tools,          this fields can store also additional bits from error status words) (TrigT1CaloTools-03-06-18)
    * Follow to the changes in xAODTrigL1Calo-00-00-55:
        * CMXCPHits_v1, CMXJetHits_v1: change size of hits0 and hits1 from uint8_t to uint32_t
          (it happens that these fields can store extra information not only thresholds)
        * CPMTower_v2, CMXCPTob_v1, CMXEtSums_v1, JetElement_v2, CMXJetTob_v1, CMXCPHits_v1:
          - change errors size from uint8_t to uint32_t (for unified usage in our tools,
             this fields can store also additional bits from error status words)
    * Tag as TrigT1CaloTools-03-06-18

2015-08-04  Alexander Mazurov <alexander.mazurov@cern.ch>
    * Merge xAOD branch
    * L1JetElementsTools, L1JEMJetTools - revert old methods for TrigT1CaloSim
    * L1CPMTools, L1CPCMXTools - use xAOD::CPMTobRoI instead of all object
    * Tag as TrigT1CaloTools-03-06-17

2015-08-03  Vasiliki Kouskoura  <Vasiliki.Kouskoura@cern.ch>
	* src/L1CPCMXTools.cxx
	- update getHits method
	- temporarily comment out formCMXCPHitsTopo from formCMXCPHits
	* Tag as TrigT1CaloTools-03-06-13-02

...
(Long ChangeLog diff - truncated)

+++ Trigger/TrigT1/TrigT1CaloTools/src/L1CPMTowerTools.cxx
@@ -277,8 +277,8 @@ void L1CPMTowerTools::makeCPMTowers(const DataVector<xAOD::TriggerTower>* trigge
       // That way if one layer is missing from TT collection we have a 0 entered already
       std::vector<uint8_t> emET(1);
       std::vector<uint8_t> hadET(1);
-      std::vector<uint8_t> emError(1);
-      std::vector<uint8_t> hadError(1);
+      std::vector<uint32_t> emError(1);
+      std::vector<uint32_t> hadError(1);
       // Both should contain same number of samples, and same peak position
       int Peak = 0;
